Tidy render warnings
This removes the fragment element ( `<>` ) which we don't need because the `<tr>` is wrapping everything. It also adds `key={i}` which gets the `i` from the `map` array method. You can google why React wants unique keys on children of arrays like this, I'm afraid I've forgotten the exact reason!